requirements.txt
mysql-connector-python==8.0.33
pyreadline==2.1